Fix balance overflow

+++ src/main/java/com/starkbank/Balance.java
@@ -12,7 +12,7 @@ import java.util.List;
 public final class Balance extends Resource {
     static ClassData data = new ClassData(Balance.class, "Balance");
 
-    public final int amount;
+    public final long amount;
     public final String currency;
     public final String updated;
 
@@ -26,11 +26,11 @@ public final class Balance extends Resource {
      * <p>
      * Attributes (return-only):
      * @param id [string, default null]: unique id returned when Balance is created. ex: "5656565656565656"
-     * @param amount [integer, default null]: current balance amount of the workspace in cents. ex: 200 (= R$ 2.00)
+     * @param amount [long, default null]: current balance amount of the workspace in cents. ex: 200 (= R$ 2.00)
      * @param currency [string, default null]: currency of the current workspace. Expect others to be added eventually. ex: "BRL"
      * @param updated [string, default null]: update datetime for the balance. ex: "2020-03-10 10:30:00.000"
      */
-    public Balance(int amount, String currency, String updated, String id) {
+    public Balance(long amount, String currency, String updated, String id) {
         super(id);
         this.amount = amount;
         this.currency = currency;
